requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Jinja2==3.1.2
requests==2.31.0
jsonschema==4.19.0
Werkzeug==2.3.7
Flask-WTF==1.1.1